Attribute amount of the model class Money fixed from long to double

+++ src/main/java/software/ulpgc/architecture/control/ExchangeMoneyCommand.java
@@ -26,7 +26,7 @@ public class ExchangeMoneyCommand implements Command {
         Money money = moneyDialog.get();
         Currency currency = currencyDialog.get();
         ExchangeRate exchangeRate = loader.load(money.currency(), currency);
-        Money result = new Money((long) (money.amount() * exchangeRate.rate()), currency);
+        Money result = new Money((double) (money.amount() * exchangeRate.rate()), currency);
         display.show(result);
     }
 }
